[DEP] Move transformations.compose to legacy and isolate uses to forecasting. (#1790)
* remove unnecessary ExponentTransform import

* compose to legacy

* compose to legacy

* import
